Titanic Disaster Survival Prediction
This project uses the Titanic's disaster dataset to make prediction on a passenger's survival based on various parameters such as age, sex, class, and family size using Random Forest Classifier, which is a famous machine learning algorithm. The model's evaluation was based on the accuracy, and the predictions were saved in a CSV file.